--- allensdk/brain_observatory/behavior/internal/__init__.py
@@ -1,2 +1,0 @@
-from allensdk.brain_observatory.behavior.internal.behavior_base import BehaviorBase  # noqa: F401
-from allensdk.brain_observatory.behavior.internal.behavior_ophys_base import BehaviorOphysBase  # noqa: F401
