+++ allensdk/brain_observatory/behavior/project_apis/abcs/__init__.py
@@ -1,0 +1,1 @@
+from allensdk.brain_observatory.behavior.project_apis.abcs.behavior_project_base import BehaviorProjectBase  # noqa: F401, E501
