+++ allensdk/brain_observatory/behavior/session_apis/abcs/__init__.py
@@ -1,0 +1,2 @@
+from allensdk.brain_observatory.behavior.session_apis.abcs.behavior_base import BehaviorBase  # noqa: F401, E501
+from allensdk.brain_observatory.behavior.session_apis.abcs.behavior_ophys_base import BehaviorOphysBase  # noqa: F401, E501
