+++ allensdk/brain_observatory/behavior/session_apis/abcs/behavior_base.py
@@ -1,6 +1,6 @@
 import abc
 
-from typing import Dict, NamedTuple
+from typing import Dict
 
 import numpy as np
 import pandas as pd
